Reduce fragility in SidePanelCoordinator (part 2 / N)
Prior to this CL, SidePanelCoordinator tracked a SidePanelEntry. As the
lifetime of SidePanelEntry is unrelated to that of SidePanelCoordinator,
this had two effects:
  * It made it impossible to distinguish between: not showing a
    side-panel entry, and showing a side-panel entry that was destroyed.
  * A proliferation of unnecessary conditionals, as the side-panel entry
    can only be destroyed in a small number of code paths.

This CL modifies SidePanelCoordinator to also track a UniqueKey. Most
logic is migrated over to track the UniqueKey. This CL fixes bug
360274920 in passing.

This CL modifies UniqueKey to track the tab handle rather than just a
tab_scoped bool. This allows SidePanelCoordinator to distinguish between
the SidePanelEntries when switching between tabs with the same
SidePanelEntry::Key.

This CL fixes a race-condition that was not seen in production code.
Production code will typically deregister a SidePanelEntry and destroy
it. However, the API for SidePanelCoordinator does not enforce this.
This means that if production code deregistered but did not destroy a
SidePanelEntry, there was a period of time (prior to the side-panel
animation finishing) where the SidePanelCoordinator was pointing to an
unregistered SidePanelEntry.

Bug: 360274920, 363743081
Change-Id: Ieaf1fd71125ffb0aa4ff1d614a7e936ef652657b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5831716
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Caroline Rising <corising@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350353}