Initialize PageDiscardingHelper in PerformanceDetectionManagerTest
The PerformanceDetectionManagerTest was silently failing when
trying to discard pages due to the PageDiscardingHelper not being
initialized in the testing suite. This CL adds an explicit check
for the PageDiscardingHelper before trying to discard a page and
initializes the helper with the graph in the test suite.

Fixed: 363264997
Change-Id: Id60617d213c339220b64a7e7731d50a4023704e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5834152
Reviewed-by: Joe Mason <joenotcharles@google.com>
Reviewed-by: Dirk Pranke <dpranke@google.com>
Commit-Queue: Steven Luong <stluong@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1350395}