Refactoring Bisect method in bisect-builds.py
This CL simplified the prefetch handling and conditions like

```
    if ((answer == 'g' and good_rev < bad_rev)
        or (answer == 'b' and bad_rev < good_rev)):
```

The new implementation reversed the rev_list to ensure rev_list[0] is
good and rev_list[-1] is bad and simply re-orders the return value. This
also fixed the issue that unknown was not working properly. Tests were
added to cover more cases.

This CL also improved the error report during run_revision. When an
error happens during run_revision, the zip_file and the tempdir will
remain on the disk while evaluating (normally AskIsGoodBuild), so that
users can debug the failure with the downloaded and unzipped archive.

Bug: 331962393
Change-Id: Ie4c19b86c0214ddabef159e44d8fe19cbb573baf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5825029
Commit-Queue: Kuan Huang <kuanhuang@chromium.org>
Reviewed-by: Sven Zheng <svenzheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350427}