Simplify FrameSerializer, step 1
FrameSerializer is only used for MHTML, but is written
as though it needs to support other formats. Because
FrameSerializer already has MHTML-specific code, this
isn't a useful design goal.

Let's simplify FrameSerializer by removing FrameSerializer::Delegate.

This will be done in two CLs for easier diffing:
1. Move FrameSerializerDelegateImpl to frame_serializer.h/cc, without
   any changes.
2. Merge FrameSerializerDelegateImpl and FrameSerializer's
   private implementation.


This CL moves code without modification. I had to remove a test and
replace it with an assertion in another test, as there is no delegate
for arbitrary URL rewriting anymore.

Bug: 363289333
Change-Id: Id39badabc6a4d5353546c9db4c680f0ce66439be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5827079
Reviewed-by: Jian Li <jianli@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Dan H <harringtond@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350429}