Move functions to ExtensionsBrowserWindow.
Move OpenOptionsPage to the ExtensionsBrowserWindow interface for
future porting and implement the ExtensionTabUtil function in terms of
that.

Do the same for BrowserSupportsTabs (now called SupportsTabs since it's
already on a "browser"-named object).

It seemed odd that this function allows a null Browser* input parameter
and the implementation that forwards it makes this more awkward. The few
callers of this function were implemented for possibly null Browser*
input so this function no longer accepts null pointers. Updates some
caller code and comments to reflect this fact and to clarify the
slightly surprising ability for GetTabById to return a null browser in
the success case.

Bug: b/361838438
Change-Id: I62c18f26d7d1f135f580a25a5bbd7e17fd199608
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5828473
Commit-Queue: Brett Wilson <brettw@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350450}