[Settings] Convert onAttachFragment to FragmentLifecycleCallbacks
Rest of onAttachFragment have been converted to
FragmentLifecycleCallbacks, so let us convert the last logging one to it
as well. Also, the new callback checks whether the fragment being
attached is the main fragment so that it does not incorrectly report
other fragments (currently there are none, but we are soon adding them).

Bug: b/356743945
Change-Id: I024d5e227570f6c1cd3a1744e1ad09f80dc6e21a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5831984
Reviewed-by: Sirisha Kavuluru <skavuluru@google.com>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350516}