Move extension loading logic to ReadAnythingService.
This CL makes two changes:
(1) It removes dead code from ReadAnythingCoordinator
(2) It moves the extension loading logic to ReadAnythingService.

The logic for (2) is moved wholesale. In the process this fixes 3 bugs
and identifies a fourth.
Bug 1: The timer to uninstall extensions was repeating when it should
have been one-shot.
Bug 2: If multiple windows are simultaneously using read anything, and
the side panel is closed on one window, the extension will be
uninstalled for the whole profile.
Bug 3: If any window is closed, the extension will be uninstalled for
the whole profile.
Bug 4: If the extension is installed but the feature is disabled (which
can happen for example if the previous run of chrome installed the
extension and then crashed), the logic to uninstall the extension will
never trigger. Documented at issue 362787711

Previously, if multiple windows were simultaneously
using read anything, then the extension could be uninstalled while the
side panel was still in use.

Bug: 355485153, 362787711
Change-Id: If94eebb46310cd5ae6f2ebc3a18fe8d7715710e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5801375
Reviewed-by: Abigail Klein <abigailbklein@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Erik Chen <erikchen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350530}