Ignore arrays declared inline for conversion to std::array
The tool struggles on these because it wants to do a textual replacement
of C array -> std::array but an std::array isn't allowed inline with the
definition of the struct. In this case we don't generate an replacement
by returning an empty Node.

We also don't work on unnamed structs (although this isn't technically
needed since it can only occur as part of the above paragraph). This
means in future if we fix the rewriter to handle inline scenarios this
should work fine, and we can fix unnamed separately.

I tried to get the class definition add a ';' and then append the
replacement but I hit the assert(len(nodes) == 2) in extract_edits.py
for

```
 # Else, parse the edge between two nodes:
 assert len(nodes) == 2
```

So for now we ignore these rewrites, I'm going to add an error locally
and log the locations that this occurs to see how often this happens in
Chrome. If it is small we can clean these up manually or if it is a lot
we can put more time in figuring out why this breaks the directed graph.

Bug: 362644557
Change-Id: I329ae090c4c5d2382fcc3f2ef69e151b8b778111
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5823071
Commit-Queue: Stephen Nusko <nuskos@chromium.org>
Auto-Submit: Stephen Nusko <nuskos@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350578}