Remove AnchorElementObserverForServiceWorker
I observed regressions in the following metrics.

- PageLoad.InteractiveTiming.InputDelay3
- WebVitals.InteractionToNextPaint2 regression

It's likely that this regression comes from
AnchorElementObserverForServiceWorker because
AnchorElementObserverForServiceWorker is a completely new code path that
is attached to the anchor event.

This CL moves the mouse event trigger of ServiceWorkerWarmUp to
AnchorElementInteractionHostImpl so that we can reduce duplicated code,
and mitigate regressions.

I don't expect behavioral change with this CL except security and
privacy check.

The behavior is checked by the following test.

All/ServiceWorkerWarmUpByPointerBrowserTest.PointeroverOrPointerdownWillWarmUpServiceWorker/*

For safer security and privacy, this CL adds RelatedActiveContents count
check. If there are other windows that might script with this frame,
this CL stops triggering the following performance optimization features.

- HttpDiskCachePrewarming
- ServiceWorkerWarmUp

See test history:
https://ci.chromium.org/ui/p/chromium/test-search?q=All%2FServiceWorkerWarmUpByPointerBrowserTest.PointeroverOrPointerdownWillWarmUpServiceWorker

Bug: 40263674
Change-Id: I730dfb858ba9e290ad17554903e1471163018bf0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5831554
Reviewed-by: Shunya Shishido <sisidovski@chromium.org>
Commit-Queue: Minoru Chikamune <chikamune@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350623}