Convert PasswordFormCache::HasPasswordForm into GetPasswordForm
Currently the cache does not expose the forms saved in it, as they're
not needed.

In order to effectively support full form filling in manual fallback we
will need access to the cached form in PasswordManualFallbackFlow, which
makes it desirable to access the form from the cache.

This CL only modifies the API and related references.

Details:

The CL achieves the switch from Has to GetPasswordForm by making use of
the PasswordFormManager::GetParsedObservedForm() method. This returns
PasswordFormManager::parsed_observed_form_ which is initialized in
PasswordFormManager::FillNow() (pageload) and never reset.

Therefore, for suggestion purposes, it is safe to assume that if there's
a PasswordForm object for a given form/field element, the new API will
find it.

Bug: 361325897
Change-Id: Ib82f7fe4f7ed6b59b845b92f61ea3fc72d0220bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5823469
Reviewed-by: Jan Keitel <jkeitel@google.com>
Reviewed-by: Christoph Schwering <schwering@google.com>
Commit-Queue: Jihad Hanna <jihadghanna@google.com>
Reviewed-by: Maria Kazinova <kazinova@google.com>
Reviewed-by: Timofey Chudakov <tchudakov@google.com>
Cr-Commit-Position: refs/heads/main@{#1350640}