[IOS] Add a lens overlay entrypoint view
This CL creates an entrypoint for the lens overlay.
The entrypoint is added as a placeholder to be displayed
when there are no badges to show.

It also addresses some visibility bugs of the placeholder view.

Bug: 361500008
Change-Id: I38b719f8d8c2df353c43f61bbecc157a55fd4b16
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5803660
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Reviewed-by: Ginny Huang <ginnyhuang@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Radu Nitescu <radunitescu@google.com>
Commit-Queue: Ameur Hosni <ameurhosni@google.com>
Cr-Commit-Position: refs/heads/main@{#1350670}