[UPMAccessLossWarning] Fix export flow crash
Checks that the file Uri is not null before trying to save passwords to
that file.
This CL also removes the "try again" option if the error dialog was
displayed during export. The reason for this is that the fragment
PasswordAccessLossExportDialogFragment would be dismissed and won't be
able to run the intent to create the file on Disk.

Bug: 364203495
Change-Id: I2051143bbfca9c6697937d61f256bb79273beb88
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5831438
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Friedrich Hauser <fhorschig@chromium.org>
Commit-Queue: Anna Tsvirchkova <atsvirchkova@google.com>
Cr-Commit-Position: refs/heads/main@{#1350705}