[ios] Mark SegmentationPlatformServiceFactory as created with Profile
Manually creating the service in DoFinalInitForServices() was a
work-around for unsafety in ServiceIsCreatedWithBrowserState().

Since the unsafety in ServiceIsCreatedWithBrowserState() has been
fixed with https://crrev.com/c/5558505, it is now the recommended
way to have the service created with the Profile.

Bug: none
Change-Id: I2151e1c9f4ec2e666d7898f529f224f56655f07f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5832036
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Federica Germinario <fedegermi@google.com>
Cr-Commit-Position: refs/heads/main@{#1350720}