[Bedtime Hours] Unhook Controller from BrowserPolicyConnector
The DeviceRestrictionScheduleController was already hooked up in a
previous CL to BrowserPolicyConnectorAsh, but since most of the code
isn't landed yet it will cause issues with upcoming CLs.

In addition, it is probably the wrong place to live and to be created
(causing lifetime issues).

Bug: b:345186543
Change-Id: I3811414ffe83feba453f0732217f6f42231629e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5832011
Reviewed-by: Igor <igorcov@chromium.org>
Commit-Queue: Ivan Šandrk <isandrk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350727}