Remove crash dumps from AutocompleteTable
As discussed offline, the NOTREACHED() statements are hit because of
I/O errors. This CL removes the reachable NOTREACHED() statements and
the crash dump emission logic.

In GetCountOfValuesContainedBetween(), 0 is returned in case of a
failure.

In AddFormFieldValueTime(), false is returned in case of a failure.
This will result in the emission of a kAddFormElements_Failure sample
to the WebDatabase.AutofillWebDataBackendImpl.OperationResult metric,
but will otherwise silently fail (meaning that no autocomplete values
get stored on submission).

Fixed: 332263206, 40260352
Change-Id: I7bc15d7b11baf18aa7eb32d4625e42b117ecdb70
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5831378
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Reviewed-by: Dominic Battré <battre@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1350756}