orderfile: Add --verify-heap to v8 flags for benchmarking
Enable --verify-heap in v8 for all benchmark runs made by
orderfile_generator. This will affect profiles, thought hopefully not by
a lot. We can restrict to arm32 and speedometer2 later if necessary.

I am assuming that adding this extra flag will help investigate the
issue on the bot for arm32 builds and soon after we will revert this
change.

I could not check that this change is working on arm32 because of my
strange configuration leading to a seemingly unrelated crash in signin.

Bug: 325104859, 349533472, 348031042
Change-Id: Id2f566410d4d82f063729043d7094e92dfde31ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5836942
Reviewed-by: Peter Wen <wnwen@chromium.org>
Commit-Queue: Peter Wen <wnwen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350804}