[DIPS] Migrate to ContentBrowserClient::IsFullCookieAccessAllowed().
Replace usage of CookieSettings::IsFullCookieAccessAllowed() with
ContentBrowserClient::IsFullCookieAccessAllowed().

Since ContentBrowserClient::IsFullCookieAccessAllowed() takes a
WebContents*, we can't call it from DIPSService, which is associated is
the BrowserContext as a whole. Instead, add an optional<bool> field to
DIPSRedirectInfo which is populated by RedirectChainDetector (which
knows its associated WebContents) before passing the redirects to the
DIPSService.

Many unit tests that call DIPSServiceImpl::HandleRedirect() or
HandleRedirectChain() had to be modified to set that field to make them
pass (the methods crash if the field is nullopt).

Bug: 358138963
Change-Id: I88169035f918c45e528a6fe5a4956752d22a3274
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5834290
Commit-Queue: Ryan Tarpine <rtarpine@chromium.org>
Reviewed-by: Anton Maliev <amaliev@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350813}