Close native Mac OS picker on page refresh
This change fixes a bug in the current code where the picker
stays active even after the page which has requested the capture
has been refreshed.

We introduce a new callback which is called after the picker is
created/opened to communicate the source_id corresponding to it
and uses that to close the picker when the page is refreshed or
another instance of the picker is opened on the same page.

Bug: 40286360
Change-Id: I835a4ecfaa7e57692340847304c0c40a13b0c666
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5833501
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Palak Agarwal <agpalak@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350814}