[iOS] Update strings in data-deletion warning dialogs on signout
This CL updates the message in these dialogs, also the sign out
confirmation button if the sign out is part of an account switch.
And passes a flag to signout action sheet coordinator to inform if it is
part of an account switch.

Also, updates the title of the AccountsTVC.

Bug: 349071954
Bug: 362708486
Change-Id: I0800a5b9327b659ce5d4e61ea81d135b3d257935
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5827545
Commit-Queue: Jood Hajeer <jood@google.com>
Reviewed-by: Arthur Milchior <arthurmilchior@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350817}