[sync] Write all theme updates to the old syncing prefs
This is as per the migration plan for the theme attributes. When
kMoveThemePrefsToSpecifics is enabled, theme will be read from
ThemeSpecifics only. However, any change will be written to both
ThemeSpecifics and to the theming prefs (which were previously used for
syncing some of the theme attributes). This is to allow new->old sync.

Bug: 356148174
Change-Id: I206ea37620200e5a5127e1d0f79833685dd7346a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5826483
Reviewed-by: Maksim Moskvitin <mmoskvitin@google.com>
Reviewed-by: Paul Adedeji <pauladedeji@google.com>
Commit-Queue: Ankush Singh <ankushkush@google.com>
Cr-Commit-Position: refs/heads/main@{#1350835}