Add ResetSettings to ProfileResetter
With the future addition of Sanitize, this new feature would need to take advantage of resetting settings. As the Sanitize feature could not sit inside the ResetSettingsHandler, this CL aims to bring in the initialization that was done inside the ResetSettingsHandler into ProfileResetter so the Sanitize feature can directly call ProfileResetter.

BUG=b:357420360, b:317898420
TEST=unit_tests --gtest_filter="*ResetSettingsHandlerTest.*"
unit_tests --gtest_filter="*ProfileResetterTest.*"

Change-Id: I80c366ae01dc86167d68e0681ce1f5d1b03629db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5761827
Reviewed-by: Wes Okuhara <wesokuhara@google.com>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Dominic Battré <battre@chromium.org>
Commit-Queue: Behnood Momenzadeh <behnoodm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350888}