Support ::cue, ::file-selector-button, and ::placeholder in getComputedStyle().
This change makes the ::cue, ::file-selector-button, and ::placeholder
pseudo-elements report correct styles in getComputedStyle() rather than
reporting styles for the element.  (It should also have similar effects
for a few other callers of the function.)

The (status:stable) PseudoElementsCorrectInGetComputedStyle flag
controls this behavior in case it needs to be reverted.

Bug: 363015176
Change-Id: Ic7c5e45685e87297bfa073c7a708b9a9c354b8ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5833587
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350908}