Change the flow of Sanitize
This CL adds the mojo API in order to kickstart the sanitize process through the Sanitize SWA. This CL also adds the sanitize button in the sanitize_initial webpage and also adds the necessary delegates.

The changes in this CL changes the flow of the feature. Basically, now the sanitize feature is called from the settings page, the initial page opens, starts the sanitize process, and once it is finished, the done page is shown, below are the screenshots of these 3 pages (two from done page):
https://screenshot.googleplex.com/BDJiq8LnRbCmAqL
https://screenshot.googleplex.com/AZrhFmTPmuDP7Qu
https://screenshot.googleplex.com/tqTure8xGXpBNcd
https://screenshot.googleplex.com/9Q44RyLv5XuEsrV

BUG=b:340724224, b:332698769, b:319446147, b:317898420

Change-Id: Ib5f569147d5c9b9b440b91d04b2158e4f2de1dc5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5649400
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Behnood Momenzadeh <behnoodm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1351018}