[Hub] Fix FAB and snackbar overlap
When the initial layout happens we should still move and a no-op will
be determined by the height being 0.

This fixes a bug where closing a tab to trigger a snackbar, leaving
the GTS and returning would cause the floating action button to not
animate out of the way.

See bug for video of fix.

Fixed: 364491503
Change-Id: I749bb9da55cc0df34c51aaa87da179e5ff54bf20
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5837840
Reviewed-by: Sky Malice <skym@chromium.org>
Auto-Submit: Calder Kitagawa <ckitagawa@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1351077}