Re-enable animation for bottom browser controls with bottom chin
This change re-enables the animations for the bottom controls, which have been disabled due to buggy interactions with the bottom chin. This change fixes some bugs with the stacker's logic for animations, where the bottom controls toolbar and mini player layers were returning incorrect `getHeight()` attributes during animations - now, the toolbar won't impact calculations with its height when it's hidden, and the mini player will continue to provide its height during its hide animation, enabling the stacker to have the proper numbers for animation calculations.

Another bug in the stacker logic has been fixed, with an instant reposition calculation after changing layers / changing the height of the controls only happens when the change is not animated. `#onBottomControlsHeightChanged()` is called before changes flow through the native code, so the browser controls are still operating with somewhat stale values, and this can interfere during animations with the later calls to `#onControlsOffsetChanged()`. Note that `#onControlsOffsetChanged()` is only called when the change is animated, so the reposition call in `#onBottomControlsHeightChanged()` is needed when the change isn't animated.

The necessary C++ changes are tracked in crrev.com/c/5809844.

No toolbar:
https://screencast.googleplex.com/cast/NjMyMDUzMzk0Njc2MTIxNnxmN2FkNjQwZC1hNA

With toolbar (thin gap is due to an unrelated detail of the mini player implementation):
https://screencast.googleplex.com/cast/NTUwOTI2NTc5NTA1NTYxNnxhN2Q2YThhOS0zNQ

Bug: 359539294
Change-Id: Ic24210a6d59887a56d6aad7ab4b609139160f022
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5815050
Reviewed-by: Ian Wells <iwells@chromium.org>
Reviewed-by: Wenyu Fu <wenyufu@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Charles Hager <clhager@google.com>
Cr-Commit-Position: refs/heads/main@{#1351105}